requirements/dependencies.txt
aiohttp==3.8.4
typing-extensions==4.5.0
python-dotenv